+++ lib/KibanaConfig.rb
